modified:   src/front/js/layout.js 	renamed:    src/front/js/pages/infoProducto.js -> src/front/js/pages/infoproducto.js
